+++ js/src/modals/Shapeshift/AwaitingExchangeStep/awaitingExchangeStep.js
@@ -17,7 +17,7 @@
 import React, { Component, PropTypes } from 'react';
 import Value from '../Value';
 
-import styles from '../fundAccount.css';
+import styles from '../shapeshift.css';
 
 export default class AwaitingExchangeStep extends Component {
   static propTypes = {
