+++ js/src/modals/Shapeshift/CompletedStep/index.js
