+++ js/src/modals/Shapeshift/Price/index.js
