+++ js/src/modals/Shapeshift/Price/price.js
@@ -16,26 +16,33 @@
 
 import React, { Component, PropTypes } from 'react';
 
-import styles from '../fundAccount.css';
+import Value from '../Value';
+import styles from '../shapeshift.css';
 
-export default class ErrorStep extends Component {
+export default class Price extends Component {
   static propTypes = {
-    error: PropTypes.shape({
-      fatal: PropTypes.bool,
-      message: PropTypes.string.isRequired
-    }).isRequired
+    coinSymbol: PropTypes.string.isRequired,
+    price: PropTypes.shape({
+      rate: PropTypes.number.isRequired,
+      minimum: PropTypes.number.isRequired,
+      limit: PropTypes.number.isRequired
+    })
   }
 
   render () {
-    const { error } = this.props;
+    const { coinSymbol, price } = this.props;
+
+    if (!price) {
+      return null;
+    }
 
     return (
-      <div className={ styles.body }>
-        <div className={ styles.info }>
-          The funds shifting via <a href='https://shapeshift.io' target='_blank'>ShapeShift.io</a> failed with a fatal error on the exchange. The error message received from the exchange is as follow:
+      <div className={ styles.price }>
+        <div>
+          <Value amount={ 1 } symbol={ coinSymbol } /> = <Value amount={ price.rate } />
         </div>
-        <div className={ styles.error }>
-          { error.message }
+        <div>
+          (<Value amount={ price.minimum } symbol={ coinSymbol } /> minimum, <Value amount={ price.limit } symbol={ coinSymbol } /> maximum)
         </div>
       </div>
     );
