+++ js/src/modals/Shapeshift/Value/index.js
