+++ js/src/modals/Shapeshift/Value/value.css
