+++ js/src/modals/Shapeshift/Value/value.js
