+++ js/src/modals/index.js
@@ -17,13 +17,13 @@
 import AddAddress from './AddAddress';
 import CreateAccount from './CreateAccount';
 import FirstRun from './FirstRun';
-import FundAccount from './FundAccount';
+import Shapeshift from './Shapeshift';
 import Transfer from './Transfer';
 
 export {
   AddAddress,
   CreateAccount,
   FirstRun,
-  FundAccount,
+  Shapeshift,
   Transfer
 };
